requirements.txt
commentjson
python-dotenv
requests